Move types around to fix ROOT error
This keeps ROOT from having to parse as much of the Celeritas
implementation.

```
In file included from input_line_9:3:
In file included from /Users/seth/Code/celeritas/src/celeritas/io/ImportData.hh:11:
In file included from /Users/seth/Code/celeritas/src/celeritas/inp/Grid.hh:13:
In file included from /Users/seth/Code/celeritas/src/corecel/grid/SplineDerivCalculator.hh:17:
In file included from /Users/seth/Code/celeritas/src/corecel/grid/UniformGrid.hh:13:
In file included from /Users/seth/Code/celeritas/src/corecel/grid/UniformGridData.hh:12:
In file included from /Users/seth/Code/celeritas/src/corecel/data/Collection.hh:15:
In file included from /Users/seth/Code/celeritas/src/corecel/data/ObserverPtr.hh:12:
/Users/seth/Code/celeritas/src/corecel/math/Algorithms.hh:763:12: error: use of undeclared identifier '__sincospif'
    return CELER_SINCOS_MANGLED(sincospif)(a, s, c);
           ^
/Users/seth/Code/celeritas/src/corecel/math/Algorithms.hh:720:9: note: expanded from macro 'CELER_SINCOS_MANGLED'
        CELER_CONCAT(CELERITAS_SINCOSPI_PREFIX, FUNC)
        ^
/Users/seth/Code/celeritas/src/corecel/math/Algorithms.hh:718:40: note: expanded from macro 'CELER_CONCAT'
                                       ^
/Users/seth/Code/celeritas/src/corecel/math/Algorithms.hh:717:45: note: expanded from macro 'CELER_CONCAT_IMPL'
                                            ^
<scratch space>:83:1: note: expanded from here
__sincospif
^
```
