+++ src/corecel/Types.cc
@@ -24,16 +24,6 @@ char const* to_cstring(MemSpace value)
 
 //---------------------------------------------------------------------------//
 /*!
- * Get a string corresponding to an interpolation.
- */
-char const* to_cstring(Interp value)
-{
-    static EnumStringMapper<Interp> const to_cstring_impl{"linear", "log"};
-    return to_cstring_impl(value);
-}
-
-//---------------------------------------------------------------------------//
-/*!
  * Get a string corresponding to a unit system.
  */
 char const* to_cstring(UnitSystem value)
