+++ src/corecel/Types.hh
@@ -79,29 +79,6 @@ enum class UnitSystem
 };
 
 //---------------------------------------------------------------------------//
-//! Interpolation type
-enum class Interp
-{
-    linear,
-    log,
-    size_
-};
-
-//---------------------------------------------------------------------------//
-/*!
- * Which of two bounding points, faces, energies, etc.
- *
- * Here, lo/hi can correspond to left/right, back/front, bottom/top. It's used
- * for the two points in a bounding box.
- */
-enum class Bound : unsigned char
-{
-    lo,
-    hi,
-    size_
-};
-
-//---------------------------------------------------------------------------//
 //!@{
 //! \name Convenience typedefs for params and states.
 
@@ -153,12 +130,6 @@ char const* to_cstring(UnitSystem);
 
 // Get a unit system corresponding to a string
 UnitSystem to_unit_system(std::string const& s);
-
-//! Convert Bound enum value to int
-CELER_CONSTEXPR_FUNCTION int to_int(Bound b)
-{
-    return static_cast<int>(b);
-}
 
 //---------------------------------------------------------------------------//
 }  // namespace celeritas
